review feedback, boiling all of index.ts down
