+++ packages/mds-db/events.ts
@@ -352,3 +352,9 @@ export async function readEventsWithTelemetry({
       : null
   }))
 }
+
+// TODO way too slow to be useful -- move into mds-cache
+export async function getMostRecentEventByProvider(): Promise<{ provider_id: UUID; max: number }[]> {
+  const sql = `select provider_id, max(recorded) from ${schema.TABLE.events} group by provider_id`
+  return makeReadOnlyQuery(sql)
+}
